Refactored models into interfaces
