+++ Interfaces/ListItemParams.ts
@@ -1,4 +1,4 @@
-export class ListItemParams {
+export interface ListItemParams {
   startBid: number;
   buyNow: number;
   duration: "1 Hour" | "3 Hour" | "6 Hour" | "12 Hours" | "1 Day" | "3 Days";
