+++ Interfaces/LoggerParams.ts
@@ -1,0 +1,3 @@
+export interface LoggerParams {
+  type: "console" | "file";
+}
