+++ Interfaces/LoginParams.ts
@@ -1,4 +1,4 @@
-export default class LoginParams {
+export interface LoginParams {
   email: string;
   password: string;
   token: string;
