--- Models/LoggerParams.ts
@@ -1,3 +1,0 @@
-export default class LoggerParams {
-  type: "console" | "file";
-}
