+++ index.ts
@@ -1,6 +1,6 @@
-export * from "./Models/ListItemParams";
-export * from "./Models/SearchPlayerParams";
-export * from "./Models/LoginParams";
+export * from "./Interfaces/ListItemParams";
+export * from "./Interfaces/SearchPlayerParams";
+export * from "./Interfaces/LoginParams";
 
 export * from "./Framework/App";
 export * from "./Interfaces/ILoginPage";
